moved the folders so that it is now in the clean architecture format
