+++ src/main/java/core/utils/TokenUtilities.java
@@ -1,4 +1,4 @@
-package utils;
+package core.utils;
 import java.util.*;
 
 /**
