+++ src/main/java/core/utils/fp/ThrowingFunction.java
@@ -1,4 +1,4 @@
-package utils.fp;
+package core.utils.fp;
 @FunctionalInterface
 public interface ThrowingFunction<T, R, E extends Exception> {
     R apply(T param) throws E;
