+++ src/main/java/core/utils/fp/ThrowingSupplier.java
@@ -1,4 +1,4 @@
-package utils.fp;
+package core.utils.fp;
 @FunctionalInterface
 public interface ThrowingSupplier<R, E extends Exception> {
     R supply() throws E;
