+++ src/main/java/domain/models/core/EventLoop.java
@@ -1,4 +1,4 @@
-package domain.eventloop;
+package domain.models.core;
 
 /**
  * This shall serve as the base for implementing an event loop. An event loop
