+++ src/main/java/domain/models/taskmanager/Deadline.java
@@ -1,4 +1,4 @@
-package domain.features.taskmanager.models;
+package domain.models.taskmanager;
 import core.exceptions.InvalidArgumentException;
 import core.utils.Pair;
 import core.utils.TokenUtilities;
