+++ src/main/java/domain/models/taskmanager/ToDo.java
@@ -1,4 +1,4 @@
-package domain.features.taskmanager.models;
+package domain.models.taskmanager;
 
 import core.exceptions.InvalidArgumentException;
