+++ src/main/java/domain/usecases/EchoUsecase.java
@@ -1,11 +1,11 @@
-package domain.features;
+package domain.usecases;
 
-import domain.eventloop.Executable;
-import domain.eventloop.ExecutableRegisterable;
-import domain.eventloop.ExitStatus;
-import domain.eventloop.NestableExecutableObject;
+import domain.models.core.Executable;
+import domain.models.core.ExecutableRegisterable;
+import domain.models.core.ExitStatus;
+import domain.models.core.NestableExecutableObject;
 
-public class Echo implements Executable, ExecutableRegisterable {
+public class EchoUsecase implements Executable, ExecutableRegisterable {
     @Override
     public ExitStatus execute(String[] tokens) {
         System.out.println(String.join(" ", tokens));
