+++ src/main/java/domain/usecases/UnknownCommandUsecase.java
@@ -1,10 +1,10 @@
-package domain.features;
-import domain.eventloop.Executable;
-import domain.eventloop.ExecutableRegisterable;
-import domain.eventloop.ExitStatus;
-import domain.eventloop.NestableExecutableObject;
+package domain.usecases;
+import domain.models.core.Executable;
+import domain.models.core.ExecutableRegisterable;
+import domain.models.core.ExitStatus;
+import domain.models.core.NestableExecutableObject;
 
-public class UnknownCommand implements Executable, ExecutableRegisterable {
+public class UnknownCommandUsecase implements Executable, ExecutableRegisterable {
     @Override
     public ExitStatus execute(String[] tokens) {
         System.out.println("☹ OOPS!!! I'm sorry, but I don't know what that " +
